\name{pbmclapply}
\alias{pbmclapply}
\title{Tracking the progress of mclapply with progress bar}
\description{
\code{pbmclapply} is a wrapper around the \code{mclapply} function. It adds a progress bar to \code{mclapply} function.
}
\usage{
pbmclapply(X, FUN, ...)
pbmclapply(X, FUN, mc.cores = 3)
pbmclapply(X, FUN, mc.style = 3)
}
\arguments{

}
\examples{
hello()
}
